requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
discord.py==2.4.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Cors==5.0.0
frozendict==2.4.6
frozenlist==1.5.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.67.0rc1
html5lib==1.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
multidict==6.1.0
multitasking==0.0.11
numpy==2.2.2
oauth2client==4.1.3
openpyxl==3.1.5
pandas==2.2.3
pandas-datareader==0.10.0
peewee==3.17.9
platformdirs==4.3.6
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pygame==2.6.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
rsa==4.9
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
webencodings==0.5.1
Werkzeug==3.1.3
yarl==1.18.3
yfinance==0.2.52
